svelte: always extract argTYpes, even without parameters.docs
