--- code/renderers/svelte/src/docs/extractComponentDescription.ts
@@ -1,10 +1,0 @@
-type Component = any;
-
-export function extractComponentDescription(component?: Component): string {
-  if (!component) {
-    return '';
-  }
-
-  const { __docgen = {} } = component;
-  return __docgen.description;
-}
