+++ code/renderers/svelte/src/extractComponentDescription.test.ts
