+++ code/renderers/svelte/src/extractComponentDescription.ts
@@ -1,0 +1,4 @@
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function extractComponentDescription(component?: any): string {
+  return component?.__docgen?.description || '';
+}
